pyproject.toml
[tool.poetry]
name = "project-nba"
version = "0.1.0"
description = ""
authors = ["stevencozine <cozineii@gmail.com>"]
readme = "README.md"

packages = [
    { include = "project_nba" }
]

[tool.poetry.dependencies]
python = "~3.11"
pandas = "^2.1.2"
google-cloud-bigquery = "^3.12.0"
apache-beam = {extras = ["gcp"], version = "^2.51.0"}
pyhocon = "^0.3.60"
scikit-learn = "^1.3.2"
nba-api = "^1.3.1"
bs4 = "^0.0.1"
lxml = "^4.9.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
